requirements.txt
fastapi
uvicorn
python-multipart
jinja2
google-genai
qdrant-client
requests
fastembed
python-dotenv 